Change method to take a pointer to a double instead of a float.

+++ include/hps_event/SvtTrack.h
@@ -130,7 +130,7 @@ class SvtTrack : public TObject {
          *
          * @parm position The extrapolated track position at the Ecal
          */
-        void setPositionAtEcal(const float* position);
+        void setPositionAtEcal(const double* position);
 
         /**
          * Set the track type.  For more details, see {@link StrategyType} and
